use int instead float

+++ database/migrations/2024_11_29_221320_create_boards_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('description')->nullable();
             $table->string('emoji')->nullable();
             $table->boolean('is_public')->default(false);
-            $table->float('order')->default(0.0)->index();
+            $table->integer('order')->default(0)->index();
             $table->foreignUuid('user_id')->constrained()->cascadeOnDelete();
             $table->timestamps();
         });
